requirements.txt
fastapi
httpx
python-dotenv
tenacity
uvicorn
aiohttp
